disable test cleanup for production mode - just in case

+++ t/99_rest.t
@@ -2,11 +2,15 @@ use Mojo::Base -strict;
 use Test::More;
 use Test::Mojo;
 
-`rm test-backups/*sql`;
-`rm backups/*dat`;
-`rm backups/*sql`;
 
 
+my $anyone = Test::Mojo->new('BibSpace');
+if($anyone->app->mode ne 'production'){	
+	`rm backups/*dat`;
+	`rm backups/*sql`;
+	`rm test-backups/*dat`;
+	`rm test-backups/*sql`;
+}
 
 ok(1);
 done_testing();
